sets up MySQL connection & refactors entity to use lombok & unit tests refactoring

+++ src/main/java/com/asofte/sucursal/repositories/SucursalRepository.java
@@ -3,7 +3,7 @@ package com.asofte.sucursal.repositories;
 import com.asofte.sucursal.domain.Sucursal;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SucursalRepository extends CrudRepository<Sucursal, Integer> {
+public interface SucursalRepository extends CrudRepository<Sucursal, Long> {
 
     Sucursal findByDireccion(String direccion);
 }
